videoAI/videosdk-deepgram-voice-agent/requirements.txt
python-dotenv==1.0.1
deepgram-sdk==3.4.0
elevenlabs==1.9.0
groq==0.11.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
httpx>=0.24.0
fastapi>=0.104.0
uvicorn>=0.24.0